requirements.txt
dj-config-url==0.1.1
dj-database-url==0.4.2
Django==1.11
django-file-resubmit==0.5.1
gunicorn==19.7.0
Pillow==5.0.0
psycopg2==2.7.3.2
pytz==2017.3
whitenoise==3.3.1
django-autocomplete-light==3.2.10
django-choices == 1.6.0
django-summernote == 0.8.8.6